First step in refactoring code to support more TDCs.
