+++ src/multiharp/mhlib_wrapper_stub.rs
